refactor Targets partials and remove redundant usings
