+++ Bullseye/Internal/DictionaryExtensions.cs
@@ -3,9 +3,7 @@ namespace Bullseye.Internal
     using System;
     using System.Collections.Generic;
     using System.Diagnostics;
-    using System.IO;
     using System.Linq;
-    using System.Reflection;
     using System.Text;
     using System.Threading.Tasks;
     using static Color;
